backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
babel==2.6.0
backcall==0.1.0           # via ipython
decorator==4.3.2          # via ipython, traitlets
django-braces==1.13.0     # via django-model-controller
django-bulk-update==2.2.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2  # via django-model-controller
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.4
django-filter==2.1.0
django-model-controller==0.2.4
django-test-without-migrations==0.6
django==2.1.5
djangorestframework-jwt==1.11.0
djangorestframework==3.9.1
gunicorn==19.9.0
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.2.0            # via ipdb
jedi==0.13.2              # via ipython
parso==0.3.2              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.4.1
prompt-toolkit==2.0.8     # via ipython
psycopg2-binary==2.7.7
ptyprocess==0.6.0         # via pexpect
pygments==2.3.1           # via ipython
pyjwt==1.7.1              # via djangorestframework-jwt
pytz==2018.9              # via babel, django
six==1.12.0               # via django-extensions, prompt-toolkit, traitlets
sqlparse==0.2.4           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
